requirements.txt
Werkzeug==2.2.2
Flask==2.2.2
SQLAlchemy==2.0.16
gunicorn==20.1.0
pyyaml==6.0
